Add kernel version check (close #185)
_NOTE: You will need to satisfy kernel requirements for Docker._ It's
currently recommended to use 3.10, although 3.8 works as well. We have
done some additional testing on Ubuntu Preciese, and found that the
oldest kernel that Ubuntu ships and weave happily runs on is 3.5,
presumably some some of the issues had been fix since Docker docs were
last updated. This is not to say that kernel version is good one to use,
we do recommend to use the latest, but just in case if you have any
specific reason why you cannot upgrade, then the 3.5 is the oldest we
have tried.