refactor: move handleGossip from LocalConnection to Router which is a better place for it to live, since that is where the gossip channel map resides.
The test code can now call that function instead of having a near
replica, so we have less test code and test more of the real code.

We also treat failure to find the gossip channel as a real failure
(which ends up closing the connection) since we generally expect peers
with the same protocol version to have the same gossip channels.