tear down connections that have no (initial) udp connectivity Prior to the introduction of gossip, a tcp connection would not see regular communication until it had been marked as 'established', because the topology communication scheme would periodically send 'FetchAll' only after that. So in the absence of udp connectivity we'd time out. The timeout could of course occur for other reasons, namely disrupted tcp connectivity, but the combination of a timeout and 'not established' pointed to a lack of udp connectivity as a likely cause. So that's what we checked for.
In the new gossip-based topology communication, however, gossip (of
anything, not just topology) occurs as soon as we have tcp
connectivity. Therefore we will no longer encounter a timeout when
just udp connectivity is disrupted. Which makes our check invalid.

So instead we introduce a one-shot timer, created on handshake
completion and cancelled when we mark the connection as
'established'. When the timer fires, and the connection still hasn't
been established, we shut down the connection with an error indicating
lack of udp connectivity.