Don't force -d on docker run
This makes it impossible for CoreOS/SystemD users to run containers as a service that utilize Weave with Restart=always and other flags and makes it where you have to use Type=oneshot and RemainAfterExit=yes, in your SystemD/Fleet unit.  This seems like a strange assumption to make on behalf of the user.