simplify connection logic slightly There is no need to remove entries from cm.targets that are in ourConnectedTargets. None of these targets will be in validTargets (since addTarget filters anything in ourConnectedTargets), so they will be removed a few lines further down anyway...
Unless the target is in the 'attempting' state. But that is ok, and in
fact desirable. Any such targets indicate connection attempts that are
in progress, which is something we display in `weave status` and want
to be accurate. Eventually the connection attempt will either succeed
or fail, and the cm.targets entry will be removed on the next run of
connectToTargets.