Set the MTU of the overlay network
We need to set the MTU of the overlay network accurately, because with
fast datapath we can't inject the ICMP frag-needed packets to cause a
lower path MTU.  Later on we can refine this to try to try to
automatically detect a good cluster-wide MTU value.  But for now we
default to a conservative value that should work on almost all local
networks, with support for manually overriding it where a higher MTU
gives better performance.